+++ types/beatmap/v3/custom/chroma.ts
@@ -10,7 +10,7 @@ import {
     ShaderKeywords,
     ShaderType,
 } from '../../shared/custom/chroma.ts';
-import { IHeckCustomEventDataBase } from './heck.ts';
+import { IHeckBase } from './heck.ts';
 import { LooseAutocomplete } from '../../../utils.ts';
 import { Vector2, Vector3 } from '../../../vector.ts';
 
@@ -182,18 +182,17 @@ export interface IChromaEventZoom extends ICustomDataBase {
 }
 
 /** AnimateComponent interface for Chroma Custom Event. */
-export interface IChromaCustomEventDataAnimateTrack extends IHeckCustomEventDataBase {
+export interface IChromaCustomEventDataAnimateTrack extends Required<IHeckBase> {
     color?: string | ColorArray | ColorPointDefinition[];
 }
 
 /** AnimateComponent interface for Chroma Custom Event. */
-export interface IChromaCustomEventDataAssignPathAnimation extends IHeckCustomEventDataBase {
+export interface IChromaCustomEventDataAssignPathAnimation extends Required<IHeckBase> {
     color?: string | ColorArray | ColorPointDefinition[];
 }
 
 /** AnimateComponent interface for Chroma Custom Event. */
-export interface IChromaCustomEventDataAnimateComponent extends IHeckCustomEventDataBase {
-    track: string;
+export interface IChromaCustomEventDataAnimateComponent extends Required<IHeckBase> {
     duration: number;
     easing?: string;
     BloomFogEnvironment?: {